refactor: cli module structures was updated to model and Adapter Command trait was added
