+++ src/cli/model/mod.rs
@@ -4,5 +4,10 @@
  *  License: MIT (See the LICENSE file in the repository root directory)
  */
 
-pub mod model;
-pub mod opt;
+pub mod check_command;
+pub mod cmp_command;
+pub mod example_command;
+pub mod output_command;
+pub mod setup_command;
+pub mod stress_command;
+pub mod traits;
