+++ src/controllers/cmd_example.rs
@@ -7,7 +7,7 @@
 use exitfailure::ExitFailure;
 
 use crate::{
-    cli::structures::ExampleCommand,
+    cli::model::example_command::ExampleCommand,
     views::example::{
         show_examples_check_cmd, show_examples_cmp_cmd, show_examples_output_cmd,
         show_examples_setup_config_cmd, show_examples_stress_cmd,
