+++ src/language/mod.rs
@@ -1,4 +1,11 @@
+/*
+ *  Quick Test: CLI for stress testing in competitive programming
+ *  Copyright (C) 2021-present / Luis Miguel Báez
+ *  License: MIT (See the LICENSE file in the repository root directory)
+ */
+
 pub mod extension;
+pub mod get_language;
 pub mod json;
 pub mod language_handler;
 pub mod traits;
